chore: re-org merge and corresponding tests (#18456)
This PR reorganizes the merge protocol code to better reflect the
architecture and improve code organization.

  ### Changes

  1. **Moved merge protocol to goblin module**
- Relocated `merge_prover.{hpp,cpp}` and `merge_verifier.{hpp,cpp}` from
`ultra_honk/` to `goblin/`
- The merge protocol is part of Goblin, so this placement better
reflects the architecture
- Separated interface from implementation by moving inline code from
headers to .cpp files
     - This improves compilation times when only implementations change

  2. **Extracted merge tests to dedicated file**
     - Created `goblin/merge.test.cpp` with all merge-specific tests
     - Removed merge tests from `mega_honk.test.cpp`
- Removed goblin dependency from ultra_honk tests (cleaner module
boundaries)

  3. **Fixed module dependencies**
     - Added goblin dependency to `stdlib_merge_verifier` module
     - Updated all include paths across the codebase

---------
