+++ barretenberg/cpp/src/barretenberg/stdlib/merge_verifier/CMakeLists.txt
@@ -1,5 +1,6 @@
 barretenberg_module(
     stdlib_merge_verifier
+    goblin
     transcript
     stdlib_poseidon2
     stdlib_circuit_builders
